# server/docs/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: SafeVideo KYC API
  description: |
    SafeVideo KYC (Know Your Customer) システムのAPIドキュメント
    
    ## 認証
    このAPIは2種類の認証方式をサポートしています：
    - **JWT認証**: `Authorization: Bearer <token>`
    - **Firebase認証**: `Firebase-Token: <token>`
    
    ## レート制限
    APIエンドポイントには以下のレート制限が適用されます：
    - デフォルト: 15分間に100リクエスト
    - バッチAPI: 15分間に10リクエスト
    - 検索API: 1分間に30リクエスト
    
    ## エラーレスポンス
    すべてのエラーレスポンスは統一された形式で返されます。
  version: 1.0.0
  contact:
    name: SafeVideo Team
    email: dev@safevideo.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 開発サーバー
  - url: https://api.safevideo.com/v1
    description: 本番サーバー
paths:
...
  /webhooks/{id}:
    get:
      tags:
        - Webhooks
      summary: Webhook詳細取得
      security:
        - bearerAuth: []
        - firebaseAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Webhook詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
    put:
      tags:
        - Webhooks
      summary: Webhook更新
      security:
        - bearerAuth: []
        - firebaseAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookConfig'
      responses:
        '200':
          description: Webhook更新完了
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
    delete:
      tags:
        - Webhooks
      summary: Webhook削除
      security:
        - bearerAuth: []
        - firebaseAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Webhook削除完了
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    firebaseAuth:
      type: apiKey
      in: header
      name: Firebase-Token
  schemas:
...
    WebhookConfig:
      type: object
      properties:
        name:
          type: string
          description: Webhook名
        url:
          type: string
          format: uri
          description: 送信先URL
        events:
          type: array
          items:
            type: string
          description: 監視するイベント
        secret:
          type: string
          description: 署名用シークレット
        active:
          type: boolean
          default: true
          description: 有効フラグ
        headers:
          type: object
          description: カスタムヘッダー
        retryCount:
          type: integer
          minimum: 0
          maximum: 5
          default: 3
          description: リトライ回数
        timeout:
          type: integer
          minimum: 1000
          maximum: 30000
          default: 5000
          description: タイムアウト（ミリ秒）
      required:
        - name
        - url
        - events
    Webhook:
      allOf:
        - $ref: '#/components/schemas/WebhookConfig'
        - type: object
          properties:
            id:
              type: integer
            userId:
              type: integer
            status:
              type: string
              enum: [active, inactive, error]
            lastTriggered:
              type: string
              format: date-time
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            statistics:
              type: object
              properties:
                totalSent:
                  type: integer
                successCount:
                  type: integer
                failureCount:
                  type: integer
                lastSuccess:
                  type: string
                  format: date-time
                lastFailure:
                  type: string
                  format: date-time
...